api: Rename api->trezorapi
