+++ trezorapi/bridge.go
@@ -1,4 +1,4 @@
-package api
+package trezorapi
 
 import (
 	"bytes"
