+++ trezorapi/functions.go
@@ -1,4 +1,4 @@
-package api
+package trezorapi
 
 import (
 	"context"
